+++ tests/test_timeseries.py
@@ -56,7 +56,7 @@ def test_arithmetic_operations():
 
 def test_hdf(tempdir):
     """Test saving/loading with HDF5."""
-    data = np.random.random((10, 10, 10, 10))
+    data = np.random.random((10, 10, 10, 10), dtype=np.float64)
     dims = ('time', 'x', 'y', 'z')
     coords = {label: np.linspace(0, 1, 10) for label in dims}
     rate = 1
